requirements.txt
altgraph==0.17.2
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
distlib==0.3.4
Django==4.0.5
django-filter==21.1
djangorestframework==3.13.1
filelock==3.7.1
flake8==4.0.1
future==0.18.2
idna==3.3
Markdown==3.3.7
mccabe==0.6.1
pefile==2021.9.3
pipenv==2022.5.2
platformdirs==2.5.2
pycodestyle==2.8.0
pyflakes==2.4.0
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.5
pytz==2022.1
pywin32-ctypes==0.2.0
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
